shopify_ai/requirements.txt
mesop
gunicorn
google-auth
google-cloud-bigquery
google-generativeai
faker
sqlparse
pandas
tabulate
db-dtypes
